Preserve growth paths with focused adapters
Split broad command and brush surfaces into feature-oriented files, document the repo architecture contract, add CI, and keep compatibility fixes small while preserving existing package imports.

Constraint: Dragonfly library consumers should keep the same import paths and command behavior.

Rejected: Move packages under pkg/ | Would add import-path noise without improving boundaries.

Rejected: Config-file dependency for stable strings | Runtime configuration need is not established.

Confidence: high

Scope-risk: moderate

Directive: Keep framework adapters thin and split files by feature before responsibilities accumulate.

Tested: go test ./...; go test -race ./...; go vet ./...; golangci-lint run ./...

Not-tested: Manual in-game Dragonfly smoke test.
