+++ act/doc.go
@@ -1,0 +1,6 @@
+// Package act registers legacy brush actions.
+//
+// It is loaded for compatibility with the legacy brush package. New brush
+// behavior should generally be implemented in editbrush and shared edit
+// primitives instead.
+package act
