+++ doc.go
@@ -1,0 +1,7 @@
+// Package we wires WorldEdit commands, session state, palette support, and
+// item-bound brushes into a Dragonfly player handler.
+//
+// The package intentionally keeps the Dragonfly adapter at the edge: block
+// editing primitives live in edit, reversible history lives in history, and
+// per-player state lives in session. Servers generally only need NewHandler.
+package we
