+++ edit/clipboard.go
@@ -2,7 +2,6 @@ package edit
 
 import (
 	"fmt"
-	"math/rand"
 
 	"github.com/df-mc/dragonfly/server/block/cube"
 	"github.com/df-mc/dragonfly/server/world"
@@ -10,15 +9,23 @@ import (
 	"github.com/df-mc/we/history"
 )
 
+// Clipboard is a captured selection used by copy, cut, paste, and schematic IO.
+//
+// OriginDir is the player's facing at copy time so paste can rotate to match
+// the new facing. Entries store offsets relative to that origin.
 type Clipboard struct {
 	OriginDir cube.Direction
 	Entries   []bufferEntry
 }
 
+// CopySelection captures area into a Clipboard. Offsets are stored relative to origin.
+// If only is true, mask filters which blocks are kept; otherwise every block (including air) is copied.
 func CopySelection(tx *world.Tx, area geo.Area, origin cube.Pos, dir cube.Direction, mask BlockMask, only bool) *Clipboard {
 	return &Clipboard{OriginDir: dir, Entries: copyArea(tx, area, origin, mask, !only)}
 }
 
+// PasteClipboard writes cb at origin, rotating around the Y axis to match dir.
+// Returns an error if the clipboard is empty. When noAir is true, air entries are skipped.
 func PasteClipboard(tx *world.Tx, cb *Clipboard, origin cube.Pos, dir cube.Direction, noAir bool, batch *history.Batch) error {
 	if cb == nil || len(cb.Entries) == 0 {
 		return fmt.Errorf("clipboard is empty")
@@ -52,8 +59,4 @@ func rotateY(pos cube.Pos, turns int) cube.Pos {
 		pos = cube.Pos{-pos[2], pos[1], pos[0]}
 	}
 	return pos
-}
-
-func randomBlock(blocks []world.Block) world.Block {
-	return ChooseBlock(blocks, rand.New(rand.NewSource(1)))
 }
