+++ geo/area.go
@@ -2,6 +2,7 @@ package geo
 
 import (
 	"fmt"
+
 	"github.com/df-mc/dragonfly/server/block/cube"
 )
