+++ geo/doc.go
@@ -1,0 +1,5 @@
+// Package geo contains geometry primitives shared by edit operations.
+//
+// Areas are inclusive cuboids and are intentionally independent from command,
+// player, form, and item concerns.
+package geo
