+++ palette/blocks.go
@@ -2,10 +2,10 @@ package palette
 
 import (
 	"encoding/json"
-	"fmt"
+	"io"
 
 	"github.com/df-mc/dragonfly/server/world"
-	"io"
+	"github.com/df-mc/we/parse"
 )
 
 // Blocks is a Palette that exists out of a slice of world.Block. It is a static palette in the sense that the
@@ -21,13 +21,13 @@ func NewBlocks(b []world.Block) Blocks {
 
 // Read reads a Blocks palette from an io.Reader.
 func Read(r io.Reader) (Blocks, error) {
-	var states []blockState
+	var states []parse.BlockState
 	if err := json.NewDecoder(r).Decode(&states); err != nil {
 		return Blocks{}, err
 	}
 	blocks := make([]world.Block, 0, len(states))
 	for _, state := range states {
-		b, err := blockFromState(state)
+		b, err := parse.BlockFromState(state)
 		if err != nil {
 			return Blocks{}, err
 		}
@@ -38,9 +38,9 @@ func Read(r io.Reader) (Blocks, error) {
 
 // Write writes a Blocks palette to an io.Writer.
 func (b Blocks) Write(w io.Writer) error {
-	states := make([]blockState, 0, len(b.b))
+	states := make([]parse.BlockState, 0, len(b.b))
 	for _, bl := range b.b {
-		states = append(states, stateOfBlock(bl))
+		states = append(states, parse.StateOfBlock(bl))
 	}
 	return json.NewEncoder(w).Encode(states)
 }
@@ -48,45 +48,4 @@ func (b Blocks) Write(w io.Writer) error {
 // Blocks returns all world.Block passed to the NewBlocks function upon creation of the palette.
 func (b Blocks) Blocks(_ *world.Tx) []world.Block {
 	return b.b
-}
-
-type blockState struct {
-	Name       string         `json:"name"`
-	Properties map[string]any `json:"properties,omitempty"`
-}
-
-func stateOfBlock(b world.Block) blockState {
-	if b == nil {
-		if air, ok := world.BlockByName("minecraft:air", nil); ok {
-			b = air
-		}
-	}
-	name, props := b.EncodeBlock()
-	if len(props) == 0 {
-		props = nil
-	}
-	return blockState{Name: name, Properties: props}
-}
-
-func blockFromState(s blockState) (world.Block, error) {
-	props := normaliseProps(s.Properties)
-	if b, ok := world.BlockByName(s.Name, props); ok {
-		return b, nil
-	}
-	return nil, fmt.Errorf("unknown block state %s", s.Name)
-}
-
-func normaliseProps(props map[string]any) map[string]any {
-	if len(props) == 0 {
-		return nil
-	}
-	out := make(map[string]any, len(props))
-	for k, v := range props {
-		if n, ok := v.(float64); ok && n == float64(int32(n)) {
-			out[k] = int32(n)
-			continue
-		}
-		out[k] = v
-	}
-	return out
 }
