+++ palette/doc.go
@@ -1,0 +1,5 @@
+// Package palette contains legacy palette selection and storage adapters.
+//
+// Palette block-state serialisation delegates to parse so persisted block
+// identity stays consistent with schematics and brush configuration.
+package palette
