+++ parse/doc.go
@@ -1,0 +1,6 @@
+// Package parse converts user-facing block, mask, and persisted block-state
+// strings into Dragonfly world values.
+//
+// It is shared by command, palette, schematic, and brush adapters so block
+// identity and JSON property normalisation stay consistent.
+package parse
